refactor(demo): move assets to a dedicated sub-folder
This avoids mixing TypeScript sources, CSS files and BPMN diagrams, to make things clearer.
